DesktopConfiguration: shared mode
This teaches the DesktopConfiguration command about a new option:
shared.  When in this mode, a shared state means that's there one set of
desktops (defined via the `DesktopName` command), but these desks are
global to all monitors, yet monitors may independently selected which
desktop to show at that time.

If a request to move to a particular desktop matches one already shown
on a different monitors, these two desks are exchanged between those
monitors.

Fixes #260